producer_kafka/requirements.txt
kafka-python-ng==2.2.3
Faker==24.11.0
python-dotenv==1.0.1
